refactor fetch behaviour
